Cleanup code structure slightly
